+++ src/build-engine/engine-ts/index.ts
@@ -17,5 +17,7 @@ export async function buildTsEngine (options: buildEngine.Options): Promise<buil
         exports: {},
         chunkAliases: { cc: './system-cc.js' },
         hasCriticalWarns: false,
+        chunkDepGraph: {},
+        assetDepGraph: {},
     };
 }