+++ test/stats-query/constant-manager.test.ts
@@ -25,7 +25,7 @@ test("exportStaticConstants", () => {
     cm.exportStaticConstants({
       mode: "PREVIEW",
       platform: "WECHAT",
-      flags: { DEBUG: false, SERVER_MODE: true, FORCE_BANNING_BULLET_WASM: true },
+      flags: { DEBUG: false, SERVER_MODE: true, FORCE_BANNING_BULLET_WASM: true, CULL_ASM_JS_MODULE: false },
     })
   ).toMatchSnapshot();
 });
@@ -35,7 +35,7 @@ test("exportDynamicConstants", () => {
     cm.exportDynamicConstants({
       mode: "BUILD",
       platform: "WECHAT",
-      flags: { DEBUG: true, SERVER_MODE: true, FORCE_BANNING_BULLET_WASM: true },
+      flags: { DEBUG: true, SERVER_MODE: true, FORCE_BANNING_BULLET_WASM: true, CULL_ASM_JS_MODULE: true },
     })
   ).toMatchSnapshot();
 });
@@ -44,7 +44,7 @@ test("genBuildTimeConstants", () => {
   const result = cm.genBuildTimeConstants({
     mode: "TEST",
     platform: "NATIVE",
-    flags: { DEBUG: false, SERVER_MODE: true, FORCE_BANNING_BULLET_WASM: true },
+    flags: { DEBUG: false, SERVER_MODE: true, FORCE_BANNING_BULLET_WASM: true, CULL_ASM_JS_MODULE: true },
   });
   expect(result).toMatchSnapshot();
 });