+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # @cocos/ccbuild
 
+## 1.1.11
+
+### Patch Changes
+
+- 58a3a35: feat: support buildEngine.enumerateAllDependents()
+
+## 1.1.10
+
+### Patch Changes
+
+- e3c3221: fix: culling bullet wasm on wechat platform
+
+## 1.1.9
+
+### Patch Changes
+
+- f7fe86e: feat: support constant CULL_ASM_JS_MODULE
+
 ## 1.1.8
 
 ### Patch Changes